+++ release_notes.md
@@ -152,3 +152,5 @@ First release into griffon artifacts portal.
 ### 0.8
 1. Allow nested property in finders.
 1. Allow user to specify custom generator for scaffolding by using `generate-all -generator=package.custom.generator`.
+1. Allow user to add scaffolding configurations in `Config.groovy`.
+1. Add `griffon.simplejpa.scaffolding.auto` to automatically run scaffolding before compiling classes.